% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{aggregation_fun}
\alias{aggregation_fun}
\title{Saldae aggregation function}
\usage{
aggregation_fun(aggregation_metric = NULL)
}
\arguments{
\item{aggregation_metric}{Maximum, Sum, Average, Minimum , Median}
}
\description{
Saldae aggregation function
}
\author{
Farid Azouaou
}
