% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection_module.R
\name{anomaly_detection_nnegh}
\alias{anomaly_detection_nnegh}
\title{Saldae Anomaly Detection wrapper}
\usage{
anomaly_detection_nnegh(
  tisefka = NULL,
  anomaly_mode = "auto",
  target_ts = NULL
)
}
\arguments{
\item{tisefka}{tible including target variable and date information}
}
\value{
tibble with anomaly information
}
\description{
anomaly detection module for for multiple data
}
\author{
Farid Azouaou
}
