% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection_module.R
\name{anomaly_detection_yiwen}
\alias{anomaly_detection_yiwen}
\title{Saldae Anomaly Detection}
\usage{
anomaly_detection_yiwen(tisefka = NULL, anomaly_mode = "auto")
}
\arguments{
\item{tisefka}{tibble with two variable date and target variable}
}
\value{
tibble with anomaly information
}
\description{
anomaly detection module for one single time serie
}
\author{
Farid Azouaou
}
