% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection_pool.R
\name{anomaly_to_DT_insight}
\alias{anomaly_to_DT_insight}
\title{Anomaly Detection Pool (View)}
\usage{
anomaly_to_DT_insight(anomalies_set = NULL)
}
\arguments{
\item{anomalies_set}{a list of dataframes containing anomaly detection results}
}
\value{
a DT table
}
\description{
Anomaly Detection Pool (View)
}
\author{
Farid Azouaou
}
