% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemmu_rate_module.R
\name{basic_growth_rate}
\alias{basic_growth_rate}
\title{basic growth rate}
\usage{
basic_growth_rate(x, asurif = 1)
}
\arguments{
\item{x}{numeric vector}

\item{asurif}{differentiation lag}
}
\value{
numeric vector containing differentiated values
}
\description{
basic growth rate
}
\author{
Farid Azouaou
}
