% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{data_aggregation_f}
\alias{data_aggregation_f}
\title{Data Aggregation based on a given time unit and grouping pattern}
\usage{
data_aggregation_f(
  tisefka = NULL,
  time_unit = NULL,
  base_unit = NULL,
  target_ts = NULL,
  sdukkel = NULL
)
}
\arguments{
\item{tisefka}{data frame including date variable}

\item{time_unit}{time unit of the aggregated data}

\item{base_unit}{time unit of the raw data}

\item{sdukkel}{group by a discrete variable}

\item{aggregation_metric}{aggregation metric (sum mean max min)}
}
\value{
list containing aggregated data
}
\description{
Data Aggregation based on a given time unit and grouping pattern
}
\author{
Farid Azouaou
}
