% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{data_exloration_aqerru}
\alias{data_exloration_aqerru}
\title{Data Exploration Main wrapper function}
\usage{
data_exloration_aqerru(
  tisefka = NULL,
  sdukkel = NULL,
  sdukkel_ih = NULL,
  aggregation_metric = NULL,
  time_unit = NULL,
  base_unit = NULL,
  target_ts = NULL,
  ts_actions = NULL
)
}
\arguments{
\item{tisefka}{data frame including date variable}

\item{sdukkel}{group by a discrete variable}

\item{sdukkel_ih}{logical variable stating whether to group or no}

\item{aggregation_metric}{aggregation metric (sum mean max min)}

\item{time_unit}{time unit of the aggregated data}

\item{base_unit}{time unit of the raw data}

\item{tisefka_report}{exploratory report of raw data}
}
\value{
list containing aggregated data with seasonally adjusted and anomaly detection infrmation
}
\description{
Data Exploration Main wrapper function
}
\author{
Farid Azouaou
}
