% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{data_exploration_pie}
\alias{data_exploration_pie}
\title{Saldae Pie Chart}
\usage{
data_exploration_pie(
  tisefka = NULL,
  x = NULL,
  y = NULL,
  aggregation_metric = "sum",
  bgrnd_color = "black"
)
}
\arguments{
\item{tisefka}{data with date as row names}

\item{x}{x-axis variable}

\item{y}{y-axis variable}

\item{bgrnd_color}{background color}

\item{aggregation}{sum mean max}
}
\value{
pie chart plotly object
}
\description{
Saldae Pie Chart
}
\author{
Farid Azouaou
}
