% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_diagnosis.R
\name{data_quality_DT}
\alias{data_quality_DT}
\title{Saldae : data diagnosis display}
\usage{
data_quality_DT(tisefka = NULL, target = "diagnosis")
}
\arguments{
\item{target}{what to display (diagnosis or outliers)}

\item{tisekfa}{data diagnosis information}
}
\value{
DT interactive object
}
\description{
Saldae : data diagnosis display
}
\author{
Farid Azouaou
}
