% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_aggregator_module.R
\name{detect_date_auto}
\alias{detect_date_auto}
\title{Saldae time step detector}
\usage{
detect_date_auto(time_vect = NULL, n_new_dates = 1, include_start_date = FALSE)
}
\arguments{
\item{time_vect}{time vector}

\item{n_new_dates}{detect only time step or generate new data using that time step}

\item{include_start_date}{whether to includelast date value}
}
\value{
time stepoutput (hours, seconds, days , ... )
}
\description{
detect based on time vector the time step of my data (hourly, daily, weekly, monthly,...)
}
\author{
Farid Azouaou
}
