% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_aggregator_module.R
\name{detect_number_basis_units_in_upper_unit}
\alias{detect_number_basis_units_in_upper_unit}
\title{Saldae basis units}
\usage{
detect_number_basis_units_in_upper_unit(
  upper_unit = NULL,
  basis_unit = NULL,
  last_date = NULL
)
}
\arguments{
\item{upper_unit}{time unit (weeks ,seconds , days , months ,quarters)}

\item{basis_unit}{basis time unit}

\item{last_date}{ur zri-gh ara}
}
\value{
a scalar number ( 24 hours in one day)
}
\description{
calculate how many basis time units in one given unit
}
\author{
Farid Azouaou
}
