% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{detect_possible_date_var}
\alias{detect_possible_date_var}
\title{possible date variables}
\usage{
detect_possible_date_var(tisefka = NULL)
}
\arguments{
\item{tisefka}{a character containing date information}
}
\value{
idk
}
\description{
Check if the intput element is a date based on a given date format
}
\author{
Farid Azouaou
}
