% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{fren_tisefka}
\alias{fren_tisefka}
\title{Saldae arrange  data}
\usage{
fren_tisefka(tifeka = NULL)
}
\arguments{
\item{tifeka}{raw data}
}
\value{
tbl_df object
}
\description{
Saldae arrange  data
}
\author{
Farid Azouaou
}
