% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemmu_rate_module.R
\name{gemmu_detect_frequency}
\alias{gemmu_detect_frequency}
\title{Saldae Detect possible growth rates}
\usage{
gemmu_detect_frequency(base_unit = NULL, gemmu_iswi = NULL)
}
\arguments{
\item{base_unit}{time unit}

\item{gemmu_iswi}{target growth rate}
}
\value{
possible growth rates
}
\description{
based on time unit detect possible growth rates
}
\author{
Farid Azouaou
}
