% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemmu_rate_module.R
\name{gemmu_yellan_f}
\alias{gemmu_yellan_f}
\title{Detect all possible growth rates based on Time Unit}
\usage{
gemmu_yellan_f(base_unit = NULL)
}
\arguments{
\item{base_unit}{time unit}
}
\value{
a vector containing possible growth rates.
}
\description{
Detect all possible growth rates based on Time Unit
}
\author{
Farid Azouaou
}
