% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{Handson_exploration}
\alias{Handson_exploration}
\title{Reactive Data Exploration including statistics data quality and graphic output.}
\usage{
Handson_exploration(
  tisefka = NULL,
  tisefka_report = NULL,
  numeric_variables = NULL
)
}
\arguments{
\item{tisefka}{Raw data(row names as date)}

\item{tisefka_report}{Data report including outliers , statistics and quality information}

\item{numeric_variables}{the numerical variables}
}
\value{
datahandontable object including start end date , statistics outliers information and charts(boxplot,distribution and raw plot)
}
\description{
Reactive Data Exploration including statistics data quality and graphic output.
}
\author{
Farid Azouaou
}
