% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values_interpolation.R
\name{interp_na_value}
\alias{interp_na_value}
\title{Saldae missing data interpolation}
\usage{
interp_na_value(ts_x = NULL, interp_mode = "spline", logic = "na_seadec")
}
\arguments{
\item{ts_x}{time serie}

\item{interp_mode}{interpolation mode linear , spline , ..}

\item{logic}{interpolation logic}
}
\value{
a vector containing a complete time series without NAs
}
\description{
use an interpolation logic based on imputeTS package to complete values
}
\author{
Farid Azouaou
}
