% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{IsDate}
\alias{IsDate}
\title{Saldae : is Date}
\usage{
IsDate(mydate, SA_date_format)
}
\arguments{
\item{mydate}{a character containing date information}

\item{SA_date_format}{date format (YYYY/mm/dd , YY-dd-mm)}
}
\value{
idk
}
\description{
Check if the intput element is a date based on a given date format
}
\author{
Farid Azouaou
}
