% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_aggregator_module.R
\name{possible_units_for_summary}
\alias{possible_units_for_summary}
\title{Saldae possible time units}
\usage{
possible_units_for_summary(time_vect = NULL)
}
\arguments{
\item{time_vect}{time vector}
}
\value{
a vector with all possible time unit valid for aggregation (hours ==> days, weeks, months)
}
\description{
based on time we derive all possible time units on which the data can be aggregated
}
\author{
Farid Azouaou
}
