% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemmu_rate_module.R
\name{rate_n_gemmu_f}
\alias{rate_n_gemmu_f}
\title{Growth Rate}
\usage{
rate_n_gemmu_f(tisefka_report = NULL, gemmu_iswi = NULL,
  d_tirni = NULL)
}
\arguments{
\item{tisefka_report}{Data exploration report}

\item{gemmu_iswi}{target growth rate}

\item{d_tirni}{exhaustive or grouped by}
}
\value{
data frame containing growth rate
}
\description{
Growth Rate
}
\author{
Farid Azouaou
}
