% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection_pool.R
\name{SA_anomaly_charter}
\alias{SA_anomaly_charter}
\title{SA anomaly detection view}
\usage{
SA_anomaly_charter(anomaly_tisefka = NULL, target_variable = NULL)
}
\arguments{
\item{anomaly_tisefka}{single variable anomaly data}

\item{targeg_variable}{the name of the variable in question}
}
\value{
a dygraph object
}
\description{
SA anomaly detection view
}
\author{
Farid Azouaou
}
