% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{SA_date_format_yellan}
\alias{SA_date_format_yellan}
\title{Saldae available date format}
\usage{
SA_date_format_yellan(ukud = NULL)
}
\arguments{
\item{ukud}{raw data}
}
\value{
vector of date format
}
\description{
Saldae available date format
}
\author{
Farid Azouaou
}
