% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemmu_rate_module.R
\name{Saldae_growth_rate_yiwen}
\alias{Saldae_growth_rate_yiwen}
\title{Saldae Growth Rate one variable}
\usage{
Saldae_growth_rate_yiwen(
  gemmu = NULL,
  tisefka = NULL,
  asurif = NULL,
  target_variable = NULL
)
}
\arguments{
\item{tisefka_report}{Data exploration report}

\item{gemmu_iswi}{target growth rate}

\item{d_tirni}{exhaustive or grouped by}
}
\value{
data frame containing growth rate
}
\description{
calculate 3 different growth rates for one single variable
}
\author{
Farid Azouaou
}
