% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemmu_rate_module.R
\name{Saldae_rate_n_gemmu_f}
\alias{Saldae_rate_n_gemmu_f}
\title{Saldae Growth Rate}
\usage{
Saldae_rate_n_gemmu_f(
  tisefka = NULL,
  base_unit = NULL,
  target_ts = NULL,
  gemmu_iswi = NULL,
  d_tirni = NULL
)
}
\arguments{
\item{gemmu_iswi}{target growth rate}

\item{d_tirni}{exhaustive or grouped by}

\item{tisefka_report}{Data exploration report}
}
\value{
data frame containing growth rate
}
\description{
Saldae Growth Rate
}
\author{
Farid Azouaou
}
