% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{sbed_tisefka}
\alias{sbed_tisefka}
\title{Saldae prepare data}
\usage{
sbed_tisefka(
  tisefka = NULL,
  date_variable = NULL,
  SA_date_format = "YYYY-MM-DD",
  aggregation_metric = NULL,
  spread_value = NULL,
  spread_key = NULL
)
}
\arguments{
\item{tisefka}{raw data(uploaded from inputfile)}

\item{date_variable}{variable to use as date}

\item{SA_date_format}{date format}

\item{aggregation_metric}{used metric to aggregate Sum, Average, Maximum,...}

\item{spread_value}{value to spread data}

\item{spread_key}{group category to use to spread data}
}
\value{
raw data ready to explore and to analyze
}
\description{
Prepare raw data using date_variable as row names
}
\author{
Farid Azouaou
}
