% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{season_adjust_f}
\alias{season_adjust_f}
\title{Seasonal Adjustement}
\usage{
season_adjust_f(tisefka = NULL, time_unit = NULL)
}
\arguments{
\item{time_unit}{time unit used to detect frequency}

\item{ts_x}{target time series}
}
\value{
a list containing frequency and seasonally adjusted time serie
}
\description{
extract from time series the seasonal , trend and remainder components based on a time frequency that we extract from time unit
}
\author{
Farid Azouaou
}
