% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemmu_rate_module.R
\name{sekned_gemmu_f}
\alias{sekned_gemmu_f}
\title{Display Growth rate results in chart.}
\usage{
sekned_gemmu_f(gemu_tisefka = NULL)
}
\arguments{
\item{gemu_tisefka}{data frame containing growth rate}
}
\value{
plotly object
}
\description{
Display Growth rate results in chart.
}
\author{
Farid Azouaou
}
