% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{sekned_tisefka}
\alias{sekned_tisefka}
\title{Display raw data including additional insights such seasonally adjusted , outliers(anomalies) and quadratic fit}
\usage{
sekned_tisefka(tisefka_report = NULL, ts_actions = TRUE,
  graph_type = "Lines", plot_settings = NULL)
}
\arguments{
\item{tisefka_report}{list containing exploratory report}

\item{graph_type}{Graph type whether a "marker" , "line" , "marker+line", "historgram", "density"}

\item{sekned_settings:}{display settings}
}
\value{
plolty reactive object
}
\description{
Display raw data including additional insights such seasonally adjusted , outliers(anomalies) and quadratic fit
}
\author{
Farid Azouaou
}
