% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{sekned_tisefka_aggregator}
\alias{sekned_tisefka_aggregator}
\title{Display raw data including additional insights such seasonally adjusted , outliers(anomalies) and quadratic fit}
\usage{
sekned_tisefka_aggregator(
  tisefka = NULL,
  time_unit = NULL,
  aggregation = NULL,
  ts_actions = NULL,
  graph_type = "Lines",
  plot_settings = NULL
)
}
\arguments{
\item{ts_actions}{I don't remember}

\item{graph_type}{Graph type whether a "marker" , "line" , "marker+line", "historgram", "density"}

\item{plot_settings}{settings to use for plotting such as color palette}

\item{tisefka_report}{list containing exploratory report}

\item{sekned_settings}{display settings}
}
\value{
plolty reactive object
}
\description{
Display raw data including additional insights such seasonally adjusted , outliers(anomalies) and quadratic fit
}
\author{
Farid Azouaou
}
