% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{sekned_tisefka_DT}
\alias{sekned_tisefka_DT}
\title{rearrange into Data frame to read it from DT functions}
\usage{
sekned_tisefka_DT(tisefka = NULL, gzem = 3)
}
\description{
rearrange into Data frame to read it from DT functions
}
