% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{tes3a_ddurt_f}
\alias{tes3a_ddurt_f}
\title{Seasonal Adjustement}
\usage{
tes3a_ddurt_f(ts_x = NULL, tuzyat = NULL)
}
\arguments{
\item{ts_x}{target time series}

\item{tuzyat}{possible frequencies}
}
\value{
seasonally adjusted time serie and statement about frequency
}
\description{
extract from time series the seasonal , trend and remainder components based on a time frequency that we extract from time unit
}
\author{
Farid Azouaou
}
