% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploratory_module.R
\name{time_based_seasonality}
\alias{time_based_seasonality}
\title{Time based Seasonality frequencies}
\usage{
time_based_seasonality(time_unit = NULL)
}
\arguments{
\item{time_unit}{time unit used to detect frequency}
}
\value{
a vector of existing frequencies for that time unit
}
\description{
detect based on time unit possible frequencies
}
\author{
Farid Azouaou
}
