% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{tisefka_spread_yella}
\alias{tisefka_spread_yella}
\title{Saldae duplicated dates in raw data}
\usage{
tisefka_spread_yella(
  tisefka = NULL,
  date_variable = NULL,
  SA_date_format = NULL,
  upper_bound = 101
)
}
\arguments{
\item{tisefka}{raw data}

\item{date_index}{variable to use as a date}
}
\value{
logical statement
}
\description{
check if there are duplicated dates in raw data and if possible create new variables based on that
}
\author{
Farid Azouaou
}
