% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_diagnosis.R
\name{ukud_acu_asdukkel}
\alias{ukud_acu_asdukkel}
\title{Saldae Time-based grouping}
\usage{
ukud_acu_asdukkel(ukud_unit = NULL)
}
\arguments{
\item{ukud_unit}{time unit}
}
\value{
time-based features(month name, days of the week , workingday or week end)
}
\description{
possible grouping time-based variables
}
\author{
Farid Azouaou
}
