% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{ukud_tilisa_f}
\alias{ukud_tilisa_f}
\title{Saldae  Time interval}
\usage{
ukud_tilisa_f(tisefka = NULL, date_vector = NULL)
}
\arguments{
\item{tisefka}{data frame with dates as row names}
}
\value{
time intervals
}
\description{
Saldae  Time interval
}
\author{
Farid Azouaou
}
