% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sprad_module.R
\name{zuzer_tisefka}
\alias{zuzer_tisefka}
\title{Saldae prepare data: spread}
\usage{
zuzer_tisefka(tisefka = NULL, anwa = NULL, f_anwa = NULL)
}
\arguments{
\item{tisefka}{raw data(uploaded from inputfile)}

\item{anwa}{value to spread data}

\item{f_anwa}{group category to use to spread data}
}
\value{
raw data ready to explore and to analyze
}
\description{
spread Data into multiple rows and have one uniqute date vector
}
\author{
Farid Azouaou
}
