% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection_pool.R
\name{anomaly_to_DT_insight}
\alias{anomaly_to_DT_insight}
\title{SA anomaly gloal viewer (DT)}
\usage{
anomaly_to_DT_insight(anomalies_set = NULL)
}
\arguments{
\item{anomalies_set}{a list of anoamly detection data}
}
\value{
DT datatable object
}
\description{
SA anomaly gloal viewer (DT)
}
\author{
Farid Azouaou
}
